@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,41 +27,40 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  autoequal: ^0.5.1
+  bloc: ^8.1.2
+  camera: ^0.10.5+2
+  cloud_firestore: ^4.8.2
+  copy_with_extension: ^5.0.2
+  cupertino_icons: ^1.0.5
+  equatable: ^2.0.5
+  firebase_analytics: ^10.4.3
+  firebase_auth: ^4.6.3
+  firebase_core: ^2.14.0
   flutter:
     sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  bloc: ^8.0.3
-  flutter_bloc: ^8.0.1
   flutter_barcode_scanner: ^2.0.0
-  equatable: ^2.0.3
+  flutter_bloc: ^8.1.3
+  flutter_cache_manager: ^3.3.0
+  flutter_cupertino_datetime_picker: ^3.0.0
   flutter_localizations:
     sdk: flutter
-  intl: ^0.17.0
-  image_picker: ^0.8.4+5
-  flutter_cupertino_datetime_picker: ^3.0.0
-  freezed_annotation: ^1.1.0
-  path: ^1.8.0
-  path_provider: ^2.0.8
-  camera: ^0.9.4+12
-  freezed: ^1.1.1
-  flutter_cache_manager: ^3.3.0
-  jiffy: ^5.0.0
-  url_launcher: ^6.0.20
-  firebase_auth: ^3.3.9
-  firebase_core: ^1.13.1
-  cloud_firestore: ^3.1.10
-  firebase_analytics: ^9.1.2
-  flutter_login: ^3.1.0
-  shared_preferences: ^2.0.13
-  google_sign_in: ^5.2.4
+  flutter_login: ^4.1.1
+  freezed: ^2.3.5
+  freezed_annotation: ^2.2.0
+  google_sign_in: ^6.1.4
+  image_picker: ^1.0.0
+  intl: ^0.18.0
+  jiffy: ^6.2.1
+  path: ^1.8.3
+  path_provider: ^2.0.15
+  shared_preferences: ^2.1.2
+  url_launcher: ^6.1.11
 
 dev_dependencies:
-  build_runner:
-  flutter_test:
-    sdk: flutter
+  autoequal_gen: ^0.5.1
+  build_runner: ^2.1.7
+  copy_with_extension_gen: ^5.0.2
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
@@ -69,10 +68,11 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^1.0.0
+  flutter_test:
+    sdk: flutter
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
   # The following line ensures that the Material Icons font is
@@ -80,18 +80,14 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   generate: true
-
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
-
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
