@@ -1,35 +1,38 @@
 part of 'auth_cubit.dart';
 
-@immutable
-abstract class AuthState {
+@autoequal
+abstract class AuthState extends Equatable {
   const AuthState();
+
+  @override
+  List<Object?> get props => _$props;
 }
 
 class AuthInitial extends AuthState {
   const AuthInitial();
+  @override
+  List<Object?> get props => _$props;
 }
 
 class AuthLoading extends AuthState {
   const AuthLoading();
+
+  @override
+  List<Object?> get props => _$props;
 }
 
 class Authenticated extends AuthState {
   final User user;
   const Authenticated(this.user);
-
   @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-
-    return other is Authenticated && other.user == user;
-  }
-
-  @override
-  int get hashCode => user.hashCode;
+  List<Object?> get props => _$props;
 }
 
 class NotAuthenticated extends AuthState {
   const NotAuthenticated();
+
+  @override
+  List<Object?> get props => _$props;
 }
 
 class AuthError extends AuthState {
@@ -37,24 +40,26 @@ class AuthError extends AuthState {
   const AuthError(this.message);
 
   @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-
-    return other is AuthError && other.message == message;
-  }
-
-  @override
-  int get hashCode => message.hashCode;
+  List<Object?> get props => _$props;
 }
 
 class PasswordRequestSubmitted extends AuthState {
   const PasswordRequestSubmitted();
+
+  @override
+  List<Object?> get props => _$props;
 }
 
 class FirstRun extends AuthState {
   const FirstRun();
+
+  @override
+  List<Object?> get props => _$props;
 }
 
 class PersonalDataUpdated extends AuthState {
   const PersonalDataUpdated();
+
+  @override
+  List<Object?> get props => _$props;
 }
