@@ -1,12 +1,13 @@
 part of 'current_warranties_cubit.dart';
 
-@freezed
-class CurrentWarrantiesState with _$CurrentWarrantiesState {
-  const factory CurrentWarrantiesState.initial({
-    @Default([]) List<WarrantyInfo> warrantyInfoList,
-    @Default([]) List<WarrantyInfo> expiringList,
-    @Default(false) bool removeBool,
-  }) = _Initial;
-
-  const CurrentWarrantiesState._();
+@CopyWith()
+class CurrentWarrantiesState {
+  const CurrentWarrantiesState({
+    required this.warrantyInfoList,
+    required this.expiringList,
+    required this.removeBool,
+  });
+  final List<WarrantyInfo> warrantyInfoList;
+  final List<WarrantyInfo> expiringList;
+  final bool removeBool;
 }
